separate PoliteaLoading, del indicators.module.css
