+++ app/components/indicators/PoliteiaLoading/PoliteiaLoading.module.css
@@ -7,17 +7,3 @@
   width: 256px;
   height: 256px;
 }
-
-.lastLogLines {
-  font-size: 9px;
-}
-
-.loaderBarEstimation {
-  text-align: left;
-  font-size: 9px;
-}
-
-.loaderBarEstimation .bold {
-  opacity: 1;
-  font-weight: 700;
-}
